méthode GET avec paramètre dans l'URL, express et postman

+++ app.js
@@ -1,5 +1,5 @@
-const con = require('./db-config');
-const express = require('express');
+const con = require("./db-config");
+const express = require("express");
 const app = express();
 app.use(express.json());
 
@@ -7,14 +7,27 @@ const port = process.env.PORT ?? 3000;
 
 con.connect((err) => {
   if (err) {
-    console.error('error connecting: ' + err.stack);
+    console.error("error connecting: " + err.stack);
   } else {
-    console.log('connected to database with threadId :  ' + con.threadId);
+    console.log("connected to database with threadId :  " + con.threadId);
   }
 });
 
 app.get("/api/movies", (req, res) => {
-  con.query("SELECT * FROM movies", (err, result) => {
+  let sql = "SELECT * FROM movies";
+  let sqlValues = [];
+  if (req.query.color) {
+    sql += " WHERE color = ?";
+    sqlValues.push(req.query.color);
+    if ((req.query.color, req.query.duration)) {
+      sql += "  AND duration <= ?";
+      sqlValues.push(req.query.duration);
+    } else if (req.query.duration) {
+      sql += " WHERE duration <= ?";
+      sqlValues.push(req.query.duration);
+    }
+  }
+  con.query(sql, sqlValues, (err, result) => {
     if (err) {
       console.error(err);
       res.status(500).send("Error retrieving data from database");
@@ -24,13 +37,47 @@ app.get("/api/movies", (req, res) => {
   });
 });
 
+app.get("/api/movies/:moviesId", (req, res) => {
+  const { moviesId } = req.params;
+  con.query("SELECT * FROM movies WHERE id = ?", [moviesId], (err, result) => {
+    if (err) {
+      console.error(err);
+      res.status(500).send("Movie not found");
+    } else if (result.length === 0 || result[0] === undefined) {
+      res.status(404).send("Movie not found");
+    } else {
+      res.json(result);
+    }
+  });
+});
+
 app.get("/api/users", (req, res) => {
-  con.query("SELECT * FROM users", (err, result) => {
+  let sql = "SELECT * FROM users";
+  const sqlValues = [];
+  if (req.query.language) {
+    sql += " WHERE language = ?";
+    sqlValues.push(req.query.language);
+  }
+  con.query(sql, sqlValues, (err, result) => {
     if (err) {
       console.error(err);
       res.status(500).send("Error retrieving users from database");
     } else {
       res.json(result);
+    }
+  });
+});
+
+app.get("/api/users/:userId", (req, res) => {
+  const { userId } = req.params;
+  con.query("SELECT * FROM users WHERE id = ? ", [userId], (err, result) => {
+    if (err) {
+      console.error(err);
+      res.status(500).send("Error retrieving users from database");
+    } else if (result.length === 0 || result[0] === undefined) {
+      res.status(404).send("user not found");
+    } else {
+      res.json(result[0]);
     }
   });
 });
@@ -67,70 +114,62 @@ app.post("/api/users", (req, res) => {
   );
 });
 
-app.put('/api/users/:userId', (req, res) => {
+app.put("/api/users/:userId", (req, res) => {
   const { userId } = req.params;
   const userPropsToUpdate = req.body;
   con.query(
-    'UPDATE users SET ? WHERE id = ?',
+    "UPDATE users SET ? WHERE id = ?",
     [userPropsToUpdate, userId],
     (err) => {
       if (err) {
         console.log(err);
-        res.status(500).send('Error updating a user');
+        res.status(500).send("Error updating a user");
       } else {
-        res.status(200).send('User updated successfully 🎉');
+        res.status(200).send("User updated successfully 🎉");
       }
     }
   );
 });
 
-app.put('/api/movies/:moviesId', (req, res) => {
+app.put("/api/movies/:moviesId", (req, res) => {
   const { moviesId } = req.params;
   const moviePropsToUpdate = req.body;
   con.query(
-    'UPDATE movies SET ? WHERE id = ?',
+    "UPDATE movies SET ? WHERE id = ?",
     [moviePropsToUpdate, moviesId],
     (err) => {
       if (err) {
         console.log(err);
-        res.status(500).send('Error updating a user');
+        res.status(500).send("Error updating a user");
       } else {
-        res.status(200).send('User updated successfully 🎉');
+        res.status(200).send("User updated successfully 🎉");
       }
     }
   );
 });
 
-app.delete('/api/users/:id', (req, res) => {
+app.delete("/api/users/:id", (req, res) => {
   const userId = req.params.id;
-  con.query(
-    'DELETE FROM users WHERE id = ?',
-    [userId],
-    (err, result) => {
-      if (err) {
-        console.log(err);
-        res.status(500).send('😱 Error deleting an user');
-      } else {
-        res.sendStatus(204);
-      }
+  con.query("DELETE FROM users WHERE id = ?", [userId], (err, result) => {
+    if (err) {
+      console.log(err);
+      res.status(500).send("😱 Error deleting an user");
+    } else {
+      res.sendStatus(204);
     }
-  );
+  });
 });
 
-app.delete('/api/movies/:moviesId', (req, res) => {
+app.delete("/api/movies/:moviesId", (req, res) => {
   const moviesId = req.params.moviesId;
-  con.query(
-    'DELETE FROM movies WHERE id = ?',
-    [moviesId],
-    (err, result) => {
-      if (err) {
-        console.log(err);
-        res.status(500).send('😱 Error deleting an user');
-      } else {
-        res.status(204).send('Film supprimé !');
-      }
+  con.query("DELETE FROM movies WHERE id = ?", [moviesId], (err, result) => {
+    if (err) {
+      console.log(err);
+      res.status(500).send("😱 Error deleting an user");
+    } else {
+      res.status(204).send("Film supprimé !");
     }
-  );
+  });
 });
 
 app.listen(port, () => {
